Move DateRange model mixin from resources to pages
