+++ cdhweb/pages/tests/test_models.py
@@ -1,6 +1,11 @@
-from unittest.mock import patch, PropertyMock
-from cdhweb.pages.models import (DisplayUrlMixin, ExternalAttachment,
-                                 LocalAttachment, RelatedLinkType)
+from datetime import date, timedelta
+
+import pytest
+from django.core.exceptions import ValidationError
+from unittest.mock import PropertyMock, patch
+
+from cdhweb.pages.models import DateRange, DisplayUrlMixin, \
+    ExternalAttachment, LocalAttachment, RelatedLinkType
 
 
 class TestRelatedLinkType:
@@ -52,3 +57,55 @@ class TestDisplayUrlMixin:
         model.url = "http://google.com/myplace#subplace"
         assert model.display_url == "google.com/myplace"
 
+
+class TestDateRange:
+
+    def test_is_current(self):
+        # start date in past, no end date
+        span = DateRange(start_date=date.today() - timedelta(days=50))
+        assert span.is_current
+
+        # end date in future
+        span.end_date = date.today() + timedelta(days=30)
+        assert span.is_current
+
+        # end date in past
+        span.end_date = date.today() - timedelta(days=3)
+        assert not span.is_current
+
+        # end date = today, current
+        span.end_date = date.today()
+        assert span.is_current
+
+        # start date in future
+        span.start_date = date.today() + timedelta(days=3)
+        assert not span.is_current
+
+    def test_years(self):
+        # start date, no end date
+        span = DateRange(start_date=date(2016, 6, 1))
+        assert span.years == '2016–'
+        # end date same year as start
+        span.end_date = date(2016, 12, 1)
+        assert span.years == '2016'
+
+        # end date known, different year
+        span.end_date = date(2017, 12, 1)
+        assert span.years == '2016–2017'
+
+    def test_clean_fields(self):
+        with pytest.raises(ValidationError):
+            DateRange(start_date=date(1901, 1, 1),
+                      end_date=date(1900, 1, 1)).clean_fields()
+
+        # should not raise exception
+        # - end after start
+        DateRange(start_date=date(1901, 1, 1), end_date=date(1905, 1, 1)) \
+            .clean_fields()
+        # - only start date set
+        DateRange(start_date=date(1901, 1, 1)).clean_fields()
+        # exclude set
+        DateRange(start_date=date(1901, 1, 1), end_date=date(1900, 1, 1)) \
+            .clean_fields(exclude=['start_date'])
+        DateRange(start_date=date(1901, 1, 1), end_date=date(1900, 1, 1)) \
+            .clean_fields(exclude=['end_date'])
